refactor: split grpc utils.rs into focused modules
Split the 1291-line utils.rs into a utils/ directory with focused modules:
- chat_utils.rs: chat processing, tool constraints, SSE helpers
- parsers.rs: reasoning/tool parser helpers
- logprobs.rs: logprob conversion functions
- metrics.rs: route_to_endpoint, error_type_from_status
- tonic_ext.rs: TonicStatusExt/TonicResultExt traits (moved from grpc/)
- mod.rs: thin re-exports only

Move single-consumer functions to their consumer files:
- get_grpc_client_from_worker -> client_acquisition.rs
- collect_stream_responses -> response_collection.rs
