+++ model_gateway/src/routers/grpc/mod.rs
@@ -12,7 +12,6 @@ pub(crate) mod pipeline;
 pub(crate) mod proto_wrapper;
 pub(crate) mod regular;
 pub(crate) mod router; // Used by routers/factory
-pub(crate) mod tonic_ext;
 pub mod utils; // Used by routers/http and bindings/golang
 
 // Re-export for convenience
