+++ model_gateway/src/routers/grpc/utils/metrics.rs
@@ -1,0 +1,28 @@
+//! Metrics helper functions (shared by HTTP routers and gRPC pipeline).
+
+use http::StatusCode;
+
+use crate::observability::metrics::metrics_labels;
+
+/// Map route path to endpoint label for metrics
+pub(crate) fn route_to_endpoint(route: &str) -> &'static str {
+    match route {
+        "/v1/chat/completions" => metrics_labels::ENDPOINT_CHAT,
+        "/generate" => metrics_labels::ENDPOINT_GENERATE,
+        "/v1/completions" => metrics_labels::ENDPOINT_COMPLETIONS,
+        "/v1/rerank" => metrics_labels::ENDPOINT_RERANK,
+        "/v1/responses" => metrics_labels::ENDPOINT_RESPONSES,
+        _ => "other",
+    }
+}
+
+/// Map HTTP status code to error type label for metrics
+pub(crate) fn error_type_from_status(status: StatusCode) -> &'static str {
+    match status.as_u16() {
+        400 => metrics_labels::ERROR_VALIDATION,
+        404 => metrics_labels::ERROR_NO_WORKERS,
+        408 | 504 => metrics_labels::ERROR_TIMEOUT,
+        500..=599 => metrics_labels::ERROR_BACKEND,
+        _ => metrics_labels::ERROR_INTERNAL,
+    }
+}
