+++ model_gateway/src/routers/grpc/utils/tonic_ext.rs
